* configure.in (SETENVOBJ): If setenv is not present on system, 	define NEED_SETENV as well as linking in local sources.
	* defines.h: Include prototype for setenv() if NEED_SETENV defined.

	* krlogin.c (prf): Declare as void. Used only with one argument -
	get rid of pseudo-varargs behaviour.

	* krlogind.c (recvauth): Include k5-util.h for
	krb5_compat_recvauth_version() prototype. Declare sendoob() void.

	* krshd.c: Include k5-util.h for krb5_compat_recvauth_version()
	prototype.

	* login.c: Include setenv prototype if NEED_SETENV defined. Cast
	arguments to printf to match format string.

	* v4rcp.c (source): Cast argument to sprintf() to match format string.
	* krcp.c (rsource): Likewise.
	* forward.c (rd_and_store_for_creds): Likewise.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13334 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/appl/bsd/v4rcp.c
@@ -524,7 +524,7 @@ notreg:
 			}
 		}
 		(void) sprintf(buf, "C%04o %ld %s\n",
-		    (unsigned int) stb.st_mode&07777, stb.st_size, last);
+		    (unsigned int) stb.st_mode&07777, (long) stb.st_size, last);
 		kstream_write (krem, buf, strlen (buf));
 		if (response() < 0) {
 			(void) close(f);
